Fix mergeFunction
